% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_model.R
\name{add_empty_model}
\alias{add_empty_model}
\title{Plot the empty model}
\usage{
add_empty_model(object, model, ...)
}
\arguments{
\item{object}{A gg object to chain to, optionally.}

\item{model}{An empty model fit by \code{\link[=lm]{lm()}}.}

\item{...}{Additional arguments to pass to \code{\link[=gf_hline]{gf_hline()}}.}
}
\value{
A gg object.
}
\description{
The empty model is extracted by pulling out \code{\link[=b0]{b0()}} from the passed model. This expression is
passed to \code{\link[=gf_hline]{gf_hline()}} to plot the empty model. If this plot is not being chained to a prior
plot, a blank point plot is created to make sure the y-axis is informative and has the range of
the original data.
}
\keyword{internal}
