% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{check_empty_model}
\alias{check_empty_model}
\title{Determine if the fitted model is the empty/null model.}
\usage{
check_empty_model(fit)
}
\arguments{
\item{fit}{A fitted linear model to pass to supernova.}
}
\description{
Determine if the fitted model is the empty/null model.
}
\keyword{internal}
