% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{check_extract_args}
\alias{check_extract_args}
\title{Assert that the arguments to the estimate extraction functions are valid.}
\usage{
check_extract_args(all, predictor, type = 3)
}
\arguments{
\item{all}{Whether to return all the estimates (e.g. all \emph{F}-values).}

\item{predictor}{The predictor(s) to return estimates for.}

\item{type}{The type (1, 2, 3) of sums of squares to use.}
}
\description{
Assert that the arguments to the estimate extraction functions are valid.
}
\keyword{internal}
