% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_data}
\alias{class_data}
\title{Generated "class data" for exploring pairwise tests}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 105 rows and 2 columns.
}
\usage{
class_data
}
\description{
These data were generated as outcomes for "students" for three different "instructors" named
A, B, and C. The outcome have means such that C > B > A, but the difference is only clearly
significant for C > A, and borderline for the others.
}
\details{
\describe{
\item{\code{outcome}}{A hypothetical, numerical outcome of an intervention.}
\item{\code{teacher}}{Either "A", "B", or "C", associating the outcome to a teacher.}
}
}
\keyword{datasets}
