% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zargle.R
\name{compare_score_distributions}
\alias{compare_score_distributions}
\title{Kargle, Bargle, and Zargle video game score simulator.}
\usage{
compare_score_distributions(sd = 3500, mean = 35000, n = 1000, ..., .seed = 5)
}
\description{
Kargle, Bargle, and Zargle are three fictitious video games referenced in the
CourseKata Statistics and Data Science course. The games are used to introduce
the concept of the standard deviation. This function simulates scores from the
three games and allows students to explore how differences in sample size, mean,
and standard deviation affect the distribution of scores.
}
