% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{convert_lm}
\alias{convert_lm}
\title{Convert a formula and data to an \code{\link{lm}} object.}
\usage{
convert_lm(object, data)
}
\arguments{
\item{object}{A \code{\link{lm}} object, or \code{\link{formula}}.}

\item{data}{If \code{object} is a formula, the data to fit the formula to as a \code{\link{data.frame}}.}
}
\value{
An \code{\link{lm}} object.
}
\description{
Convert a formula and data to an \code{\link{lm}} object.
}
\keyword{internal}
