% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{convert_predictor}
\alias{convert_predictor}
\title{Convert a potentially complex predictor to a character vector of terms.}
\usage{
convert_predictor(predictor)
}
\arguments{
\item{predictor}{The predictor(s) to return estimates for.}
}
\value{
A character vector of terms.
}
\description{
Convert a potentially complex predictor to a character vector of terms.
}
\keyword{internal}
