% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{coursekata_attach}
\alias{coursekata_attach}
\title{Attach the CourseKata course packages}
\usage{
coursekata_attach(startup = FALSE)
}
\arguments{
\item{startup}{Is this being run at start-up?}
}
\value{
An object with info about which course packages are installed and attached.
}
\description{
Attach the CourseKata course packages
}
\examples{
coursekata_attach()
}
