% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{coursekata_attach_message}
\alias{coursekata_attach_message}
\title{Information about CourseKata packages.}
\usage{
coursekata_attach_message(pkgs)
}
\arguments{
\item{pkgs}{A character vector of packages being loaded.}
}
\value{
A coursekata_attachments object, also of class data.frame with a row for each course
package and a column for each of the \code{package} name, \code{version}, and whether it is currently
\code{attached}.
}
\description{
Information about CourseKata packages.
}
\keyword{internal}
