% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_packages.R
\name{coursekata_attached}
\alias{coursekata_attached}
\title{List all currently attached CourseKata course packages}
\usage{
coursekata_attached()
}
\value{
A character vector of the course packages that have been attached.
}
\description{
List all currently attached CourseKata course packages
}
\keyword{internal}
