% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{coursekata_attachments}
\alias{coursekata_attachments}
\title{Information about CourseKata packages.}
\usage{
coursekata_attachments(startup = FALSE)
}
\arguments{
\item{startup}{Is this being run at start-up?}
}
\value{
A coursekata_attachments object, also of class data.frame with a row for each course
package and a column for each of the \code{package} name, \code{version}, and whether it is
currently \code{attached}.
}
\description{
Information about CourseKata packages.
}
\keyword{internal}
