% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{coursekata_attachments_message}
\alias{coursekata_attachments_message}
\title{Pretty console printing when attaching packages}
\usage{
coursekata_attachments_message(pkg_info)
}
\arguments{
\item{pkg_info}{The package information to print.}
}
\value{
A string that has been formatted to print to console.
}
\description{
Pretty console printing when attaching packages
}
\keyword{internal}
