% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_packages.R
\name{coursekata_detached}
\alias{coursekata_detached}
\title{List all currently NOT attached CourseKata course packages}
\usage{
coursekata_detached()
}
\value{
A character vector of the course packages that are not attached.
}
\description{
List all currently NOT attached CourseKata course packages
}
\keyword{internal}
