% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_install.R
\name{coursekata_install}
\alias{coursekata_install}
\alias{coursekata_update}
\title{Install or update all CourseKata packages.}
\usage{
coursekata_install(...)

coursekata_update(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\value{
The state of all the packages after any updates have been performed.
}
\description{
Install or update all CourseKata packages.
}
