% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_packages.R
\name{coursekata_packages}
\alias{coursekata_packages}
\title{List all CourseKata course packages}
\usage{
coursekata_packages(core_only = FALSE)
}
\arguments{
\item{core_only}{Optionally suppress the data packages.}
}
\value{
A character vector of the names of the course packages.
}
\description{
List all CourseKata course packages
}
\examples{
coursekata_packages()
}
