% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_coursekata.R
\name{coursekata_palettes}
\alias{coursekata_palettes}
\title{The color palettes used in our theme system}
\usage{
coursekata_palettes()
}
\value{
A named list of the different color palettes used is returned with the exception of the
continuous palette. The continuous palette can be accessed by calling \code{viridisLite::viridis(n)}
where \code{n} is the number of colors to retrieve.
}
\description{
There are four main color palettes used in our theme system:
\itemize{
\item \strong{single-color}: for plots with one color, the CourseKata blue from the website is used
\item \strong{eight-color}: for plots with up to eight colors, the colors from
\code{ggtheme::scale_color_colorblind} are used, but the orders of some of the colors are swapped
\item \strong{fifteen-color}: when a broader palette of up to 15 colors is needed, a different
colorblind-friendly palette is used, from \link{http://mkweb.bcgsc.ca/biovis2012/}.
\item \strong{continuous}: for plots with continuous colorspaces, the \code{\link[viridisLite:viridis]{viridisLite::viridis()}} palette
is used
}
}
