% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_extraction.R
\name{extract_stat}
\alias{extract_stat}
\title{Extract a statistic from a supernova table and name the values}
\usage{
extract_stat(fit, type, stat, predictor = character(0))
}
\arguments{
\item{fit}{A fitted linear model to pass to supernova.}

\item{type}{The type (1, 2, 3) of sums of squares to use.}

\item{stat}{The statistic from the supernova (as named in the \code{supernova(...)$tbl}).}

\item{predictor}{Optionally specify which terms to return (instead of all of them).}
}
\description{
Extract a statistic from a supernova table and name the values
}
\keyword{internal}
