% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_data}
\alias{game_data}
\title{Simulated math game data.}
\format{
A data frame with 105 observations on the following 2 variables:
\describe{
\item{\code{game}}{The game the student was randomly assigned to, coded as "A", "B", or "C".}
\item{\code{outcome}}{Each student's score on the outcome test.}
}
}
\usage{
game_data
}
\description{
The simulated results of a small study comparing the effectiveness of three different computer-
based math games in a sample of 105 fifth-grade students. All three games focused on the same
topic and had identical learning goals, and none of the students had any prior knowledge of the
topic.
}
\keyword{datasets}
