% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_packages.R
\name{is_attached}
\alias{is_attached}
\title{Check if packages are attached}
\usage{
is_attached(package)
}
\arguments{
\item{package}{Character vector of the names of the packages to check.}
}
\value{
Logical vector indicating whether the packages are attached or not.
}
\description{
Check if packages are attached
}
\keyword{internal}
