% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{library_location}
\alias{library_location}
\title{Determine which library a package was loaded from}
\usage{
library_location(package)
}
\arguments{
\item{package}{The package name string.}
}
\value{
If the package is loaded, the location it was loaded from, else NULL
}
\description{
Determine which library a package was loaded from
}
\keyword{internal}
