% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_coursekata.R
\name{load_coursekata_themes}
\alias{load_coursekata_themes}
\title{Utility function for loading all themes.}
\usage{
load_coursekata_themes()
}
\description{
This function is called at package start-up and should rarely be needed by the user. The
exception is when the user has called \code{\link[=restore_default_themes]{restore_default_themes()}} and wants to go back to the
CourseKata look and feel. When run, this function sets the CourseKata color palettes
\code{\link{coursekata_palettes}}, sets the default theme to \code{\link{theme_coursekata}}, and tweaks some default
settings for specific plots. To restore the original \code{ggplot2} settings, run
\code{\link[=restore_default_themes]{restore_default_themes()}}.
}
\seealso{
coursekata_palettes theme_coursekata scale_discrete_coursekata restore_default_themes
}
