% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit.R
\name{m_predictor}
\alias{m_predictor}
\title{Get the predictor variable from a model.}
\usage{
m_predictor(model)
}
\arguments{
\item{model}{An \code{\link{lm}} model (or anything compatible with \code{\link[=formula]{formula()}}).}
}
\value{
A string with the name of the predictor variable.
}
\description{
Get the predictor variable from a model.
}
\keyword{internal}
