% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_attach.R
\name{package_version}
\alias{package_version}
\title{Get a package's version}
\usage{
package_version(package)
}
\arguments{
\item{package}{The package name string.}
}
\value{
The version of the package as a character string. If the package is already loaded, this
is pulled from the library the package was loaded from, else the default library location.
}
\description{
Get a package's version
}
\keyword{internal}
