% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{A modified form of the \code{\link[palmerpenguins:penguins]{palmerpenguins::penguins}} data set.}
\format{
A data frame with 333 observations on the following 7 variables:
\describe{
\item{\code{species}}{The species of penguin, coded as "Adelie", "Chinstrap", or "Gentoo".}
\item{\code{gentoo}}{Whether the penguin is a Gentoo penguin (1) or not (0).}
\item{\code{body_mass_kg}}{The mass of the penguin's body, in kilograms.}
\item{\code{flipper_length_m}}{The length of the penguin's flipper, in m.}
\item{\code{bill_length_cm}}{The length of the penguin's bill, in cm.}
\item{\code{female}}{Whether the penguin is female (1) or not (0).}
\item{\code{island}}{The island where the penguin was observed, coded as "Biscoe", "Dream", or "Torgersen".}
}
}
\usage{
penguins
}
\description{
The modifications are to select only a subset of the variables, and convert some of the units.
}
\keyword{datasets}
