% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_install}
\alias{pkg_install}
\title{Install packages, but intercept for fivethirtyeightdata}
\usage{
pkg_install(pkg)
}
\arguments{
\item{pkg}{A character vector of the packages to install.}
}
\description{
fivethirtyeight data needs to be loaded from an alternate repo.
}
\keyword{internal}
