% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_is_installed}
\alias{pkg_is_installed}
\title{check if packages are installed.}
\usage{
pkg_is_installed(pkgs)
}
\arguments{
\item{pkgs}{Character vector of the names of the packages to check.}
}
\value{
Logical vector indicating whether the packages are attached or not.
}
\description{
check if packages are installed.
}
\keyword{internal}
