% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_remote_version}
\alias{pkg_remote_version}
\title{Get the remote package version numbers.}
\usage{
pkg_remote_version(pkgs)
}
\arguments{
\item{pkgs}{A character vector of packages to look up.}
}
\value{
A character vector of the package versions, or NA if the package cannot be found.
}
\description{
Get the remote package version numbers.
}
\keyword{internal}
