% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_require}
\alias{pkg_require}
\title{Load the package, making sure to load from the correct location. Sometimes people have different
versions of packages installed, so if the package is currently loaded, make sure to load from the
place it was already loaded from.}
\usage{
pkg_require(pkg, do_not_ask = FALSE)
}
\description{
Load the package, making sure to load from the correct location. Sometimes people have different
versions of packages installed, so if the package is currently loaded, make sure to load from the
place it was already loaded from.
}
\keyword{internal}
