% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_version}
\alias{pkg_version}
\title{Get package versions.}
\usage{
pkg_version(pkg)
}
\arguments{
\item{pkg}{A character vector of packages to look up.}
}
\value{
A character vector of the package versions. If the package is already loaded, this is
pulled from the library the package was loaded from, else the default library location.
}
\description{
Get package versions.
}
\keyword{internal}
