% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_coursekata.R
\name{restore_default_themes}
\alias{restore_default_themes}
\title{Restore \code{ggplot2} default settings}
\usage{
restore_default_themes()
}
\description{
This function will restore all of the tweaks to themes and plotting to the original \code{ggplot2}
defaults. If you want to go back to the CourseKata look and feel, run
\code{\link[=load_coursekata_themes]{load_coursekata_themes()}}.
}
\seealso{
load_coursekata_themes
}
