% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zargle.R
\name{simulate_scores}
\alias{simulate_scores}
\title{Simulate score for one of the *argle games.}
\usage{
simulate_scores(game, n, mean, sd)
}
\arguments{
\item{game}{The name of the game.}

\item{n}{The number of scores to simulate.}

\item{mean}{The mean of the scores.}

\item{sd}{The standard deviation of the scores.}
}
\value{
A data frame with the simulated scores.
}
\description{
Simulate score for one of the *argle games.
}
\keyword{internal}
