% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Smallville}
\alias{Smallville}
\title{Simulated housing data}
\format{
A data frame with 32 observations on the following 4 variables:
\describe{
\item{\code{PriceK}}{Price the home sold for (in thousands of dollars)}
\item{\code{Neighborhood}}{The neighborhood the home is in (Eastside, Downtown)}
\item{\code{HomeSizeK}}{The size of the home (in thousands of square feet)}
\item{\code{HasFireplace}}{Whether the home has a fireplace (0 = no, 1 = yes)}
}
}
\usage{
Smallville
}
\description{
These data are simulated to be similar to the Ames housing data, but with far fewer variables
and much smaller effect sizes.
}
\keyword{datasets}
