% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit.R
\name{test_fit}
\alias{test_fit}
\title{Test the fit of a model on a train and test set.}
\usage{
test_fit(model, df_train, df_test)
}
\arguments{
\item{model}{An \code{\link{lm}} model (or anything compatible with \code{\link[=formula]{formula()}}).}

\item{df_train}{A data frame with the training data.}

\item{df_test}{A data frame with the test data.}
}
\value{
A data frame with the fit statistics.
}
\description{
Test the fit of a model on a train and test set.
}
