% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tip_exp}
\alias{tip_exp}
\title{Simulated data for an experiment about smiley faces and tips}
\format{
A data frame with 44 observations on the following 3 variables.
\describe{
\item{\code{gender}}{Whether the server was \code{female} or \code{male}}
\item{\code{condition}}{Whether the check had a \verb{smiley face} or not (\code{control})}
\item{\code{tip_percent}}{The size of the tip as a percentage of the price of the meal}
}
}
\usage{
tip_exp
}
\description{
These are simulated data that are similar to the \code{TipExperiment} data. Hypothetical tables
were randomly assigned to receive checks that either included or did not include a drawing
of a smiley face, either from a male or a female server.
}
\keyword{datasets}
