% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_cols}
\alias{to_cols}
\title{Split a character into columns for terminal output}
\usage{
to_cols(strings, n_cols = 2, space_between = "       ")
}
\arguments{
\item{strings}{The strings to divide into columns.}

\item{n_cols}{The number of columns.}

\item{space_between}{What to put between the columns.}
}
\value{
A string that prints to the terminal as columns.
}
\description{
Split the string into \code{n} columns, then glue the columns together row-wise with
\code{space_between}, then glue the rows together with new line characters.
}
\keyword{internal}
