% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_model.R
\name{add_group_model}
\alias{add_group_model}
\title{Plot the group model}
\usage{
add_group_model(object, gformula, data, width = 0.3, ...)
}
\arguments{
\item{object}{A gg object to chain to, optionally.}

\item{gformula}{A formula of the shape \code{y ~ x}.}

\item{data}{The data the formula refers to.}

\item{width}{The width of the mean lines to be plotted.}

\item{...}{Additional arguments passed to \code{\link[=gf_segment]{gf_segment()}}}
}
\value{
A gg object.
}
\description{
The group model is represented by mean lines for each group in the model. The values for these
lines are extracted using the formula with the passed data, and then plotted via
\code{\link[=gf_segment]{gf_segment()}}. If this plot is not being chained to a prior plot, a blank point plot is
created to make sure the axes are informative and have the range of the original data.
}
\keyword{internal}
