% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{ask_to_install}
\alias{ask_to_install}
\title{Ask the user if they want to install packages}
\usage{
ask_to_install(pkgs)
}
\arguments{
\item{pkgs}{A character vector of packages to ask about.}
}
\value{
A logical indicating whether the use answered yes or no.
}
\description{
Ask the user if they want to install packages
}
\keyword{internal}
