% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata-package.R
\docType{package}
\name{coursekata-package}
\alias{coursekata}
\alias{coursekata-package}
\title{coursekata: Packages and Functions for CourseKata Courses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Easily install and load all packages and functions used in CourseKata courses. Aid teaching with helper functions and augment generic functions to provide cohesion between the network of packages. Learn more about CourseKata at \url{https://www.coursekata.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/UCLTALL/coursekata-r}
  \item Report bugs at \url{https://www.github.com/UCLTALL/coursekata-r/issues}
}

}
\author{
\strong{Maintainer}: Adam Blake \email{adamblake@g.ucla.edu} (\href{https://orcid.org/0000-0001-7881-8652}{ORCID})

Authors:
\itemize{
  \item Ji Son \email{json2@calstatela.edu} (\href{https://orcid.org/0000-0002-4258-4791}{ORCID})
  \item Jim Stigler \email{jstigler@gmail.com} (\href{https://orcid.org/0000-0001-6107-7827}{ORCID})
}

}
\keyword{internal}
