% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_install.R
\name{coursekata_install}
\alias{coursekata_install}
\alias{coursekata_update}
\title{Install or update all CourseKata packages.}
\usage{
coursekata_install(...)

coursekata_update(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[remotes:install_cran]{remotes::install_cran}} or \code{\link[remotes:install_github]{remotes::install_github}}
depending on whether the package appears to be from CRAN or GitHub.}
}
\value{
The state of all the packages after any updates have been performed.
}
\description{
Install or update all CourseKata packages.
}
