% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_packages.R
\name{coursekata_packages}
\alias{coursekata_packages}
\title{List all CourseKata course packages}
\usage{
coursekata_packages()
}
\value{
A data frame with three variables: the name of the package \code{package}, the
\code{version}, and whether it is currently \code{attached}.
}
\description{
List all CourseKata course packages
}
\examples{
coursekata_packages()
}
