% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{coursekata_palette}
\alias{coursekata_palette}
\title{The color palettes used in our theme system}
\usage{
coursekata_palette(indices = integer(0))
}
\arguments{
\item{indices}{The indices of the colors to pull (or all colors if no indices are given).}
}
\value{
A named list of the requested colors in the palette.
}
\description{
The color palettes used in our theme system
}
