% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{coursekata_palette_provider}
\alias{coursekata_palette_provider}
\title{Create a function that provides a colorblind palette.}
\usage{
coursekata_palette_provider()
}
\value{
A function that accepts one argument \code{n}, which is the number of colors you want to use
in the plot. This function is used by scales like \code{scale_color_discrete} to provide colorblind-
safe palettes. See \code{\link{scale_discrete_coursekata}} for more information.
}
\description{
Create a function that provides a colorblind palette.
}
