% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursekata_repos.R
\name{coursekata_repos}
\alias{coursekata_repos}
\title{Get repositories for the packages.}
\usage{
coursekata_repos(repos = getOption("repos"))
}
\arguments{
\item{repos}{Optionally set a repository character vector to augment.}
}
\value{
A set of repositories that can be used to install or update the CourseKata packages.
}
\description{
Ensures a default CRAN is set if one is not already set, and adds the repository for
fivethirtyeightdata.
}
\examples{
coursekata_repos()
}
