% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{coursekata_unload_theme}
\alias{coursekata_unload_theme}
\title{Restore \code{ggplot2} default settings}
\usage{
coursekata_unload_theme()
}
\value{
No return value, called to restore the global state of \code{ggplot2}.
}
\description{
This function will restore all of the tweaks to themes and plotting to the original \code{ggplot2}
defaults. If you want to go back to the CourseKata look and feel, run
\code{\link[=coursekata_load_theme]{coursekata_load_theme()}}.
}
\seealso{
coursekata_load_theme
}
