% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_is_installed}
\alias{pkg_is_installed}
\title{Check if packages are installed}
\usage{
pkg_is_installed(pkgs)
}
\arguments{
\item{pkgs}{Character vector of the names of the packages to check.}
}
\value{
Logical vector indicating whether the packages are installed.
}
\description{
Note: this function differs from \code{\link[rlang:is_installed]{rlang::is_installed()}} in two regards: it is quieter and will
show no messages, and it returns a vector indicating which packages are installed or not (rather
than a single Boolean value regarding the packages as a set).
}
\keyword{internal}
