% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{pkg_library_location}
\alias{pkg_library_location}
\title{Determine which libraries packages were loaded from}
\usage{
pkg_library_location(pkg)
}
\arguments{
\item{pkg}{A character vector of packages to check.}
}
\value{
A character vector of library directory paths, or NA if the package is not loaded.
}
\description{
Determine which libraries packages were loaded from
}
\keyword{internal}
