% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_discrete_coursekata}
\alias{scale_discrete_coursekata}
\title{A discrete color scale constructor with colorblind-safe palettes.}
\usage{
scale_discrete_coursekata(...)
}
\arguments{
\item{...}{Additional parameters passed on to the scale type.}
}
\value{
A discrete color scale.
}
\description{
See \code{\link[=coursekata_palette]{coursekata_palette()}} for more information.
}
\seealso{
coursekata_palette
}
