% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_coursekata}
\alias{theme_coursekata}
\title{A simple theme built on top of \code{\link[=theme_bw]{theme_bw()}}}
\usage{
theme_coursekata()
}
\value{
A gg theme object
}
\description{
The \code{coursekata} package automatically loads this theme when the package is loaded. This is in
addition to a number of other plot tweaks and option settings. To just restore the theme to the
default, you can run \code{set_theme(theme_grey)}. If you want to restore all plot related settings
and/or prevent them when loading the package, see \code{\link{coursekata_unload_theme}}.
}
\examples{
gf_boxplot(Thumb ~ RaceEthnic, data = Fingers, fill = ~RaceEthnic)
}
